Move `EndianBuf` to endianity.rs
